backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.4.0
python-dotenv==1.0.0